requirements.txt
asyncio==3.4.3
asyncpg==0.21.0
docker==4.3.1
docker-compose==1.27.4
psycopg2-binary==2.8.6

PyYAML==5.3.1

tabulate==0.8.7
PyInquirer==1.0.3

PyInquirer==1.0.3
uvloop==0.14.0
aiofiles==0.6.0
aiohttp==3.7.3
tornado==6.0.4


